+++ src/controllers/caronaController.js
@@ -1,23 +1,52 @@
 const caronaSchema = require("../models/caronaSchema");
+const jwt = require("jsonwebtoken");
+const SECRET = process.env.SECRET;
 const mongoose = require("mongoose");
 
 const getAllCaronas = async (request, response) => {
+
+    const authHeader = request.get('authorization');
+    if (!authHeader) {
+        return response.status(401).send("Não autorizado");
+    }
+
+    const token = authHeader.split(' ')[1];
+
+    jwt.verify(token, SECRET, async function (erro) {
+        if (erro) {
+            return response.status(403).send('Token invalido!');
+    }
+
     try {
         const caronas = await caronaSchema.find();
         response.status(200).json(caronas);
     } catch (error) {
         response.status(500).json({message: error.message});        
     }
+  })  
 }
 
 const createCarona = async (request, response) => {
+
+    const authHeader = request.get('authorization');
+    if (!authHeader) {
+        return response.status(401).send("Não autorizado");
+    }
+
+    const token = authHeader.split(' ')[1];
+
+    jwt.verify(token, SECRET, async function (erro) {
+        if (erro) {
+            return response.status(403).send('Token invalido!');
+    }
+
     try {
         const carona = new caronaSchema({
             anuncioCarona: request.body.anuncioCarona,
             usuarioCarona: request.body.usuarioCarona,
             _id: new mongoose.Types.ObjectId
         })
-
+        
         const novaCarona = await carona.save();
 
         response.status(200).json(
@@ -31,9 +60,23 @@ const createCarona = async (request, response) => {
     } catch (error) {
         response.status(500).json({message: error.message});        
     }
+  })
 }
 
 const updateCarona = async (request, response) => {
+
+    const authHeader = request.get('authorization');
+    if (!authHeader) {
+        return response.status(401).send("Não autorizado");
+    }
+
+    const token = authHeader.split(' ')[1];
+
+    jwt.verify(token, SECRET, async function (erro) {
+        if (erro) {
+            return response.status(403).send('Token invalido!');
+    }
+
     try {
         const caronaEncontrada = await caronaSchema.findById(request.params.id);
 
@@ -55,9 +98,23 @@ const updateCarona = async (request, response) => {
     } catch (error) {
         response.status(500).json({message: error.message});        
     }
+  })  
 }
 
 const deleteCarona = async (request, response) => {
+
+    const authHeader = request.get('authorization');
+    if (!authHeader) {
+        return response.status(401).send("Não autorizado");
+    }
+
+    const token = authHeader.split(' ')[1];
+
+    jwt.verify(token, SECRET, async function (erro) {
+        if (erro) {
+            return response.status(403).send('Token invalido!');
+    }
+
     try {
         const caronaEncontrada = await caronaSchema.findById(request.params.id);
         caronaEncontrada.delete();
@@ -73,6 +130,7 @@ const deleteCarona = async (request, response) => {
     } catch (error) {
         response.status(500).json({message: error.message});        
     }
+  })  
 }
 
 module.exports = {
